GUI changes
Implemented changes so that routes will better handle closures.
Still need to fix graph. Much better than start of today.
More visual user prompts just so the user knows what is occuring.

+++ Code/Cycle 2/ARP/DeleteRoute.java
@@ -1,3 +1,5 @@
+import javax.swing.JOptionPane;
+
 /*
  * To change this template, choose Tools | Templates
  * and open the template in the editor.
@@ -184,8 +186,15 @@ public class DeleteRoute extends javax.swing.JFrame {
 
 	private void deleteButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_deleteButtonActionPerformed
 		
+		
+		int result=JOptionPane.showConfirmDialog(this,
+				("Are you sure you want to delete "+routeBox.getSelectedItem().toString()+" ?"),"Confirmation", JOptionPane.INFORMATION_MESSAGE);
+		if(result==JOptionPane.OK_OPTION){
 		d.deleteRoute((Route) routeBox.getSelectedItem());
+		JOptionPane.showMessageDialog(this,
+				("Route "+routeBox.getSelectedItem().toString()+" has been deleted!"),"Buh bye!", JOptionPane.INFORMATION_MESSAGE);
 		routeBox.setModel(new javax.swing.DefaultComboBoxModel(d.getAllRoutes().toArray()));
+		}
 	}//GEN-LAST:event_deleteButtonActionPerformed
 
 	private void destFieldActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_destFieldActionPerformed
